@@ -10,12 +10,11 @@
     "@capacitor/filesystem": "^1.1.0",
     "@capacitor/ios": "^3.5.1",
     "@capacitor/storage": "^1.2.5",
-    "@chakra-ui/icons": "^1.1.7",
-    "@chakra-ui/react": "^1.8.5",
+    "@chakra-ui/icons": "^2.0.2",
+    "@chakra-ui/react": "^2.1.2",
     "@ctrl/tinycolor": "^3.4.0",
     "@emotion/react": "^11.8.1",
     "@emotion/styled": "^11.8.1",
-    "@radix-ui/react-tooltip": "^0.1.6",
     "@tauri-apps/api": "^1.0.0-rc.0",
     "@xstate/test": "^0.5.1",
     "flexsearch": "^0.7.21",
@@ -23,8 +22,8 @@
     "i18next": "^21.6.11",
     "jotai": "^1.6.0",
     "lodash-es": "^4.17.21",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^18.1.0",
+    "react-dom": "^18.1.0",
     "react-i18next": "^11.15.4",
     "react-query": "^3.34.15",
     "react-router-dom": "^6.3.0",
@@ -50,7 +49,7 @@
     "@types/react-virtualized": "^9.21.18",
     "@typescript-eslint/eslint-plugin": "^5.27.0",
     "@typescript-eslint/parser": "^5.27.0",
-    "@vitejs/plugin-react": "^1.0.7",
+    "@vitejs/plugin-react": "^1.3.2",
     "@vitest/ui": "^0.3.4",
     "c8": "^7.11.0",
     "concurrently": "^7.2.1",
