@@ -4,7 +4,7 @@ import 'react-virtualized/styles.css';
 
 import { ChakraProvider, useColorMode } from '@chakra-ui/react';
 import React from 'react';
-import ReactDOM from 'react-dom';
+import { createRoot } from 'react-dom/client';
 import { HashRouter } from 'react-router-dom';
 
 import App from '../App';
@@ -39,11 +39,12 @@ const Wrapper = pipeWrapper(
   }),
 );
 
-ReactDOM.render(
+const root = createRoot(document.getElementById('root'));
+
+root.render(
   <React.StrictMode>
     <Wrapper>
       <App />
     </Wrapper>
-  </React.StrictMode>,
-  document.getElementById('root')
+  </React.StrictMode>
 );
